Fixes issue #37 - synch dialog box crashing.
This fix changes the photo request url to the server to request for a resided photo. Previsiouly the dialog would hang because the client would run out of memory when processing large photos from the server.
The request to the server now includes a custom size for the picture that the server will return.