@ctumwebaze @mwangiann rapidftr/tracker#47 enter and search multi-lingual data
entering and searching multi-lingual data worked out of the box. The mingled character issues
was due to encoding data during the sync process.
When a sync request to push data to the server was constructed, the proper endcoding was not provided when
constructing the body of the request and this caused the special characters like the accent characters to get
replaced with question marks or boxes.

The solution to this was to specify the default encoding which is utf-8 on android.