rapidftr/tracker#201 @ctumwebaze @mwangiann listing potential matches
The reason why the pontential matches view was displaying a single record was because of embedding a listview in a scrollview.
This PotentialMatchesView was inheriting a FormSectionView which was inheriting a ScrollView. What this fix does is to alter the inheritance change and only make the FormSectionView (Now DefaultFormSectionView) inherit from a ScrollView and the PotentialMatchesView inherit from a LinearLayout.